+++ parallax/calibration.py
@@ -3,7 +3,7 @@
 import numpy as np
 import cv2 as cv
 from . import lib
-from .Helper import WF, HF
+from .helper import WF, HF
 
 
 imtx1 = [[1.81982227e+04, 0.00000000e+00, 2.59310865e+03],
